+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.6][] - 2022-12-01
+
 ### Added
 
 -   New component in React: `PopoverDialog`.
@@ -1656,3 +1658,5 @@ _Failed released_
 [3.0.4]: https://github.com/lumapps/design-system/tree/v3.0.4
 [unreleased]: https://github.com/lumapps/design-system/compare/v3.0.5...HEAD
 [3.0.5]: https://github.com/lumapps/design-system/tree/v3.0.5
+[unreleased]: https://github.com/lumapps/design-system/compare/v3.0.6...HEAD
+[3.0.6]: https://github.com/lumapps/design-system/tree/v3.0.6